move config utils into cmd
That is the logical place to them as they are cmd related
