+++ cmd/config/config_test.go
@@ -1,4 +1,4 @@
-package utils
+package config
 
 import (
 	. "github.com/onsi/gomega"
